feat: Extract releases files subcommand to a separate files command
